+++ terraform/cdn.tf
@@ -48,6 +48,39 @@ resource "aws_cloudfront_distribution" "bare_domain" {
   }
 }
 
+resource "aws_iam_role" "lambda_edge_assume_role" {
+  provider           = aws.iam
+  name               = "lambda_edge_assume_role"
+  assume_role_policy = data.aws_iam_policy_document.lambda_edge_assume_role.json
+
+  tags = {
+    Managed = true
+  }
+}
+
+# resource "aws_iam_role_policy" "lambda_edge_execute" {
+#   provider = aws.iam
+#   name     = "lambda_edge_execute"
+#   policy   = data.aws_iam_policy_document.lambda_edge_execute.json
+#   role     = aws_iam_role.lambda_edge_assume_role.id
+# }
+
+resource "aws_lambda_function" "inject_headers" {
+  provider         = aws.lambda
+  filename         = data.archive_file.inject_headers_lambda.output_path
+  function_name    = "inject-headers"
+  role             = aws_iam_role.lambda_edge_assume_role.arn
+  handler          = "inject_headers.handler"
+  publish          = true
+  source_code_hash = filebase64sha256(data.archive_file.inject_headers_lambda.output_path)
+  # As of 2021-03-25 Lambda@Edge doesn't support nodejs14.x.
+  runtime = "nodejs12.x"
+
+  tags = {
+    Managed = true
+  }
+}
+
 resource "aws_cloudfront_distribution" "www_domain" {
   origin {
     domain_name = aws_s3_bucket.www_domain.bucket_domain_name
@@ -58,11 +91,11 @@ resource "aws_cloudfront_distribution" "www_domain" {
     }
   }
 
+  aliases             = ["www.${var.domain}"]
   enabled             = true
   is_ipv6_enabled     = true
   default_root_object = "index.html"
-
-  aliases = ["www.${var.domain}"]
+  price_class         = "PriceClass_All"
 
   default_cache_behavior {
     allowed_methods          = ["GET", "HEAD", "OPTIONS"]
@@ -71,11 +104,14 @@ resource "aws_cloudfront_distribution" "www_domain" {
     compress                 = true
     origin_request_policy_id = data.aws_cloudfront_origin_request_policy.cors_s3.id
     target_origin_id         = "S3-www.${var.domain}"
+    viewer_protocol_policy   = "redirect-to-https"
 
-    viewer_protocol_policy = "redirect-to-https"
+    lambda_function_association {
+      event_type   = "origin-response"
+      include_body = false
+      lambda_arn   = aws_lambda_function.inject_headers.qualified_arn
+    }
   }
-
-  price_class = "PriceClass_All"
 
   restrictions {
     geo_restriction {